ref: rename oauthRoute.js to oauthRoutes.js and update imports
